move popups categories to its own folder
