+++ apps/fishing-map/features/map/popups/categories/AnnotationTooltip.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { DeckLayerPickingObject } from '@globalfishingwatch/deck-layers'
 import { selectIsGuestUser } from 'features/user/selectors/user.selectors'
-import styles from './Popup.module.css'
+import styles from '../Popup.module.css'
 
 type AnnotationTooltipProps = {
   features: DeckLayerPickingObject[]
