+++ apps/fishing-map/features/map/popups/categories/ContextLayers.tsx
@@ -7,7 +7,7 @@ import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import { selectContextAreasDataviews } from 'features/dataviews/selectors/dataviews.selectors'
 import { getDatasetTitleByDataview } from 'features/datasets/datasets.utils'
-import styles from './Popup.module.css'
+import styles from '../Popup.module.css'
 import ContextLayersRow from './ContextLayersRow'
 import { useContextInteractions } from './ContextLayers.hooks'
