+++ apps/fishing-map/features/map/popups/categories/DetectionsLayers.tsx
@@ -3,10 +3,12 @@ import { useTranslation } from 'react-i18next'
 import { Icon, Spinner } from '@globalfishingwatch/ui-components'
 import { DataviewCategory } from '@globalfishingwatch/api-types'
 import I18nNumber from 'features/i18n/i18nNumber'
-import VesselsTable, { VesselDetectionTimestamps } from 'features/map/popups/VesselsTable'
-import { SliceExtendedFourwingsDeckSublayer } from '../map.slice'
-import { getVesselsInfoConfig } from '../map.hooks'
-import styles from './Popup.module.css'
+import VesselsTable, {
+  VesselDetectionTimestamps,
+} from 'features/map/popups/categories/VesselsTable'
+import { SliceExtendedFourwingsDeckSublayer } from '../../map.slice'
+import { getVesselsInfoConfig } from '../../map.hooks'
+import styles from '../Popup.module.css'
 
 type ViirsMatchTooltipRowProps = {
   feature: SliceExtendedFourwingsDeckSublayer & { category: DataviewCategory; title?: string }
