+++ apps/fishing-map/features/map/popups/categories/VesselsTable.module.css
