+++ apps/fishing-map/features/reports/Report.tsx
@@ -55,7 +55,7 @@ import { getDatasetsReportNotSupported } from 'features/datasets/datasets.utils'
 import DatasetLabel from 'features/datasets/DatasetLabel'
 import { LAST_REPORTS_STORAGE_KEY, LastReportStorage } from 'features/reports/reports.config'
 // import { REPORT_BUFFER_GENERATOR_ID } from 'features/map/map.config'
-import { useHighlightReportArea } from 'features/map/popups/ContextLayers.hooks'
+import { useHighlightReportArea } from 'features/map/popups/categories/ContextLayers.hooks'
 import { selectIsGuestUser, selectUserData } from 'features/user/selectors/user.selectors'
 import { useFetchDataviewResources } from 'features/resources/resources.hooks'
 import { useFetchReportArea, useFetchReportVessel, useFitAreaInViewport } from './reports.hooks'
