+++ apps/fourwings-explorer/features/map/popups/ContextPopup.tsx
@@ -1,7 +1,7 @@
 import { ExtendedFeature } from '@globalfishingwatch/react-hooks'
 import { useDatasetLayers } from 'features/layers/layers.hooks'
 import { ContextApiDatasetConfig } from 'features/datasets/datasets.types'
-import styles from './Popup.module.css'
+import styles from '../Popup.module.css'
 
 type ContextPopupProps = {
   feature?: ExtendedFeature
